Global setting of test_engine and clone_db (#89)
This PR adds support to setting the `test_engine` and `clone_db` in a global fashion.

The idea is that we can set which engine to use for running the tests, and from which database to clone for each test. We currently can do that on a per-`@test_rel` basis, but we would like to support a way to globally define a default.

This has been tested with the package manager testing harness.

This PR also includes changes done by the `JuliaFormatter`.